+++ learning/katas/go/core_transforms/combine/combine_perkey/cmd/main.go
@@ -16,7 +16,7 @@
 package main
 
 import (
-	"combine_perkey/pkg/task"
+	"beam.apache.org/learning/katas/core_transforms/combine/combine_perkey/pkg/task"
 	"context"
 	"github.com/apache/beam/sdks/go/pkg/beam"
 	"github.com/apache/beam/sdks/go/pkg/beam/log"
