+++ learning/katas/go/core_transforms/combine/combine_perkey/task-info.yaml
@@ -23,13 +23,9 @@ files:
   visible: false
 - name: cmd/main.go
   visible: true
-- name: go.mod
-  visible: false
 - name: pkg/task/task.go
   visible: true
   placeholders:
   - offset: 1021
     length: 88
     placeholder_text: TODO()
-- name: go.sum
-  visible: false
