+++ learning/katas/go/core_transforms/combine/combine_perkey/task-remote-info.yaml
