+++ learning/katas/go/core_transforms/combine/lesson-info.yaml
@@ -20,4 +20,4 @@
 content:
 - simple_function
 - combinefn
-- Combine PerKey
+- combine_perkey
